refactor paths and files
