+++ build.gradle
@@ -1,0 +1,19 @@
+
+apply plugin: 'java'
+
+repositories {
+    jcenter()
+}
+
+dependencies {
+    compile 'org.slf4j:slf4j-api:1.7.24'
+    testCompile 'junit:junit:4.12'
+    testCompile "org.mockito:mockito-core:2.+"
+
+}
+
+jar {
+    manifest {
+        attributes 'Main-Class': 'BibliotecaApp'
+    }
+}
